requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
streamlit>=1.28.0
schedule>=1.2.0
pandas>=2.2.0
numpy>=1.26.0
pyupbit>=0.2.31
PyJWT>=2.8.0
python-telegram-bot>=20.5
schedule>=1.2.0
google-generativeai>=0.3.0
websocket-client>=1.6.3
cryptography>=41.0.4
aiohttp>=3.8.5
plotly>=5.15.0
